Rename files and add to doumentation
