+++ docs/source/setup_matrices.rst
@@ -7,9 +7,16 @@ Setup Matrices
 Overview:
 
 - :class:`ContactMatrix`
+- :class:`RegionMatrix`
 
 Contact Population Matrix
 *************************
 
 .. autoclass:: ContactMatrix
+  :members:
+
+Region-specific Susceptibility Matrix
+*************************************
+
+.. autoclass:: RegionMatrix
   :members:
